renames prodcucts/add to create
